RF: move readlinkabs definition to fix recursive import
